requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.assertions>=1.132.0
aws-cdk.aws_ec2>=1.132.0
aws-cdk.aws_lambda>=1.132.0